move my registrations and public registration out of registration folder
