+++ src/pages/my_registrations/DeleteRegistrationModal.tsx
