+++ src/pages/public_registration/PublicPublicationInstance.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from 'react';
-import LabelContentRow from '../../../components/LabelContentRow';
-import { JournalPublicationInstance } from '../../../types/publication_types/journalPublication.types';
+import LabelContentRow from '../../components/LabelContentRow';
+import { JournalPublicationInstance } from '../../types/publication_types/journalPublication.types';
 import { useTranslation } from 'react-i18next';
-import { DegreePublicationInstance } from '../../../types/publication_types/degreePublication.types';
-import { ReportPublicationInstance } from '../../../types/publication_types/reportPublication.types';
-import { PagesMonograph } from '../../../types/publication.types';
+import { DegreePublicationInstance } from '../../types/publication_types/degreePublication.types';
+import { ReportPublicationInstance } from '../../types/publication_types/reportPublication.types';
+import { PagesMonograph } from '../../types/publication.types';
 
 export const PublicPublicationInstanceJournal: FC<{ publicationInstance: JournalPublicationInstance }> = ({
   publicationInstance,
