+++ src/pages/public_registration/PublicRegistrationAuthors.tsx
@@ -3,12 +3,12 @@ import styled from 'styled-components';
 import { Link, IconButton } from '@material-ui/core';
 import MailOutlineIcon from '@material-ui/icons/MailOutline';
 
-import { Contributor } from '../../../types/contributor.types';
-import NormalText from '../../../components/NormalText';
-import { getDistinctContributorUnits } from '../../../utils/institutions-helpers';
+import { Contributor } from '../../types/contributor.types';
+import NormalText from '../../components/NormalText';
+import { getDistinctContributorUnits } from '../../utils/institutions-helpers';
 import OrcidLogo from '../../../resources/images/orcid_logo.svg';
-import { ORCID_BASE_URL } from '../../../utils/constants';
-import AffiliationHierarchy from '../../../components/institution/AffiliationHierarchy';
+import { ORCID_BASE_URL } from '../../utils/constants';
+import AffiliationHierarchy from '../../components/institution/AffiliationHierarchy';
 
 const StyledAuthor = styled.span`
   margin-right: 1rem;
